+++ connectors/core/cloud-connector/src/test/java/org/platformlambda/mock/TestBase.java
@@ -19,9 +19,9 @@
 package org.platformlambda.mock;
 
 import org.junit.BeforeClass;
+import org.platformlambda.core.system.AppStarter;
 import org.platformlambda.core.util.AppConfigReader;
 import org.platformlambda.core.util.Utility;
-import org.platformlambda.rest.RestServer;
 
 import java.util.concurrent.atomic.AtomicInteger;
 
@@ -36,7 +36,7 @@ public class TestBase {
             Utility util = Utility.getInstance();
             AppConfigReader config = AppConfigReader.getInstance();
             port = util.str2int(config.getProperty("server.port", "8085"));
-            RestServer.main(new String[0]);
+            AppStarter.main(new String[0]);
         }
     }
 }