+++ language-packs/language-connector/src/main/java/org/platformlambda/lang/services/LanguageRelay.java
@@ -27,7 +27,6 @@ import org.platformlambda.core.system.PostOffice;
 import org.platformlambda.core.util.CryptoApi;
 import org.platformlambda.core.websocket.common.MultipartPayload;
 import org.platformlambda.lang.websocket.server.LanguageConnector;
-import org.platformlambda.websocket.WsConfigurator;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -52,7 +51,7 @@ public class LanguageRelay implements LambdaFunction {
     private static final String COUNT = MultipartPayload.COUNT;
     private static final String TOTAL = MultipartPayload.TOTAL;
     private static final int OVERHEAD = MultipartPayload.OVERHEAD;
-    private static final int MAX_PAYLOAD_SIZE = WsConfigurator.getInstance().getMaxBinaryPayload() - OVERHEAD;
+    private static final int MAX_PAYLOAD_SIZE = 64 * 1024 - OVERHEAD;
     private static final LanguageRelay instance = new LanguageRelay();
 
     private LanguageRelay() {