+++ system/platform-core/src/main/java/org/platformlambda/core/system/AppStarter.java
@@ -19,23 +19,33 @@
 package org.platformlambda.core.system;
 
 import io.github.classgraph.ClassInfo;
+import io.vertx.core.Vertx;
+import io.vertx.core.http.HttpServerOptions;
 import org.platformlambda.core.annotations.BeforeApplication;
 import org.platformlambda.core.annotations.MainApplication;
+import org.platformlambda.core.annotations.WebSocketService;
 import org.platformlambda.core.models.EntryPoint;
+import org.platformlambda.core.models.LambdaFunction;
+import org.platformlambda.core.util.AppConfigReader;
 import org.platformlambda.core.util.Feature;
 import org.platformlambda.core.util.SimpleClassScanner;
 import org.platformlambda.core.util.Utility;
+import org.platformlambda.core.websocket.server.ActuatorServiceHandler;
+import org.platformlambda.core.websocket.server.WsRequestHandler;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import java.lang.reflect.InvocationTargetException;
 import java.util.*;
+import java.util.concurrent.ConcurrentHashMap;
+import java.util.concurrent.ConcurrentMap;
 
 public class AppStarter {
     private static final Logger log = LoggerFactory.getLogger(AppStarter.class);
+    private static final ConcurrentMap<String, LambdaFunction> lambdas = new ConcurrentHashMap<>();
 
     private static final int MAX_SEQ = 999;
     private static boolean loaded = false;
-    private static boolean webapp = false;
     private static String[] args = new String[0];
 
     public static void main(String[] args) {
@@ -43,19 +53,13 @@ public class AppStarter {
             loaded = true;
             AppStarter.args = args;
             AppStarter begin = new AppStarter();
+            // run "BeforeApplication" modules
             begin.doApps(args, false);
-            /*
-             * Do not start MainApplication(s) if the app is packaged to run as a WAR file.
-             * The web application server will start the WAR file directly.
-             */
-            if (!webapp) {
-                begin.doApps(args, true);
-            }
+            // run "MainApplication" modules
+            begin.doApps(args, true);
+            // setup websocket server if required
+            begin.startWebSocketServerIfAny();
         }
-    }
-
-    public static void setWebApp(boolean enable) {
-        AppStarter.webapp = enable;
     }
 
     public static String[] getArgs() {
@@ -92,11 +96,11 @@ public class AppStarter {
 
     private int getSequence(Class<?> cls, boolean main) {
         if (main) {
-            MainApplication before = cls.getAnnotation(MainApplication.class);
-            return Math.min(MAX_SEQ, before.sequence());
+            MainApplication mainApp = cls.getAnnotation(MainApplication.class);
+            return Math.min(MAX_SEQ, mainApp.sequence());
         } else {
-            BeforeApplication before = cls.getAnnotation(BeforeApplication.class);
-            return Math.min(MAX_SEQ, before.sequence());
+            BeforeApplication beforeApp = cls.getAnnotation(BeforeApplication.class);
+            return Math.min(MAX_SEQ, beforeApp.sequence());
         }
     }
 
@@ -130,6 +134,82 @@ public class AppStarter {
             log.error("Missing MainApplication\n\n{}\n{}\n\n",
                     "Did you forget to annotate your main module with @MainApplication that implements EntryPoint?",
                     "and ensure the package parent is defined in 'web.component.scan' of application.properties.");
+        }
+    }
+
+    private void startWebSocketServerIfAny() {
+        // find and execute optional preparation modules
+        SimpleClassScanner scanner = SimpleClassScanner.getInstance();
+        Set<String> packages = scanner.getPackages(true);
+        for (String p : packages) {
+            List<ClassInfo> services = scanner.getAnnotatedClasses(p, WebSocketService.class);
+            for (ClassInfo info : services) {
+                try {
+                    Class<?> cls = Class.forName(info.getName());
+                    if (Feature.isRequired(cls)) {
+                        WebSocketService annotation = cls.getAnnotation(WebSocketService.class);
+                        if (annotation.value().length() > 0) {
+                            if (!Utility.getInstance().validServiceName(annotation.value())) {
+                                log.error("Unable to load {} ({}) because the path is not a valid service name",
+                                        cls.getName(), annotation.value());
+                            }
+                            loadLambda(cls, annotation.namespace(), annotation.value());
+                        }
+                    } else {
+                        log.info("Skipping optional {}", cls);
+                    }
+                } catch (ClassNotFoundException e) {
+                    log.error("Class {} not found", info.getName());
+                }
+            }
+        }
+        // start websocket server
+        if (!lambdas.isEmpty()) {
+            Utility util = Utility.getInstance();
+            AppConfigReader config = AppConfigReader.getInstance();
+            int port = util.str2int(config.getProperty("websocket.server.port",
+                                    config.getProperty("server.port", "8085")));
+            if (port > 0) {
+                Vertx vertx = Vertx.vertx();
+                HttpServerOptions options = new HttpServerOptions().setTcpKeepAlive(true);
+                vertx.createHttpServer(options)
+                        .requestHandler(new ActuatorServiceHandler())
+                        .webSocketHandler(new WsRequestHandler(lambdas))
+                        .listen(port)
+                        .onSuccess(server -> {
+                            log.info("Websocket server running on port-{}", server.actualPort());
+                        })
+                        .onFailure(ex -> {
+                            log.error("Unable to start - {}", ex.getMessage());
+                            System.exit(-1);
+                        });
+            }
+        }
+    }
+
+    private void loadLambda(Class<?> cls, String namespace, String value) {
+        Utility util = Utility.getInstance();
+        List<String> parts = util.split(namespace + "/" + value, "/");
+        StringBuilder sb = new StringBuilder();
+        for (String p: parts) {
+            sb.append('/');
+            sb.append(p);
+        }
+        String path = sb.toString();
+        String wsEndpoint = path + "/{handle}";
+        try {
+            Object o = cls.getDeclaredConstructor().newInstance();
+            if (o instanceof LambdaFunction) {
+                lambdas.put(path, (LambdaFunction) o);
+                log.info("{} loaded as WEBSOCKET SERVER endpoint {}", cls.getName(), wsEndpoint);
+            } else {
+                log.error("Unable to load {} ({}) because it is not an instance of {}",
+                        cls.getName(), wsEndpoint, LambdaFunction.class.getName());
+            }
+
+        } catch (InstantiationException | IllegalAccessException | NoSuchMethodException |
+                 InvocationTargetException e) {
+            log.error("Unable to load {} ({}) - {}", cls.getName(), wsEndpoint, e.getMessage());
         }
     }