+++ system/rest-spring/src/main/java/org/platformlambda/rest/RestServer.java
@@ -34,12 +34,6 @@ public class RestServer extends SpringBootServletInitializer {
      * @param args may not be relevant when deployed as WAR
      */
     public static void main(String[] args) {
-        /*
-         * execute preparation steps if any
-         * - this allows application to do preparation such as setting environment variables,
-         *   overriding application.properties, etc.
-         */
-        AppStarter.setWebApp(true);
         AppStarter.main(args);
         // start Spring Boot
         SpringApplication.run(RestServer.class, args);